feat: 프로필 섹션 컴포넌트 구현
